docs: rename categories.json to categories.react.json
