+++ docs/app/data/categories.react.json
