--- docs/app/react/side-nav.json
@@ -1,241 +1,0 @@
-[
-  {
-    "label": "Overview",
-    "type": "heading"
-  },
-  {
-    "label": "Getting Started",
-    "route": "/react",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Loading States",
-    "route": "/react/loading-states",
-    "tag": "new",
-    "type": "route"
-  },
-  {
-    "label": "Actions",
-    "type": "heading"
-  },
-  {
-    "label": "Button",
-    "route": "/react/button",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Icon Button",
-    "route": "/react/icon-button",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Communication",
-    "type": "heading"
-  },
-  {
-    "label": "Avatar",
-    "route": "/react/avatar",
-    "tag": "new",
-    "type": "route"
-  },
-  {
-    "label": "Notification",
-    "route": "/react/notification",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Progress Bar",
-    "route": "/react/progress-bar",
-    "tag": "new",
-    "type": "route"
-  },
-  {
-    "label": "Tag",
-    "route": "/react/tags",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Tooltip",
-    "route": "/react/tooltip",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Containment",
-    "type": "heading"
-  },
-  {
-    "label": "Confirm Modal",
-    "route": "/react/confirm-modal",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Prompt Modal",
-    "route": "/react/prompt-modal",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Modal",
-    "route": "/react/modal",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Table",
-    "route": "/react/table",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Navigation",
-    "type": "heading"
-  },
-  {
-    "label": "Nav Menu",
-    "route": "/react/nav-menu",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Tabs",
-    "route": "/react/tabs",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Selection",
-    "type": "heading"
-  },
-  {
-    "label": "Drag & Drop",
-    "route": "/react/drag-n-drop",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Checkbox",
-    "route": "/react/checkbox",
-    "tag": "new",
-    "type": "route"
-  },
-  {
-    "label": "Radio",
-    "route": "/react/radio",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Select",
-    "route": "/react/select",
-    "tag": "new",
-    "type": "route"
-  },
-  {
-    "label": "Toggle",
-    "route": "/react/toggle",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Forms",
-    "type": "heading"
-  },
-  {
-    "label": "Field Message",
-    "route": "/react/field-message",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "File Uploader",
-    "route": "/react/file-uploader",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Input",
-    "route": "/react/input",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Label",
-    "route": "/react/label",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Textarea",
-    "route": "/react/textarea",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Rendering",
-    "type": "heading"
-  },
-  {
-    "label": "Show",
-    "route": "/react/show",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Portal",
-    "route": "/react/portal",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Feature Flags",
-    "route": "/react/feature-flags",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Hooks",
-    "type": "heading"
-  },
-  {
-    "label": "useModal",
-    "route": "/react/use-modal",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "useTheme",
-    "route": "/react/use-theme",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "useThemeContext",
-    "route": "/react/use-theme-context",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "useToggle",
-    "route": "/react/use-toggle",
-    "tag": "",
-    "type": "route"
-  },
-  {
-    "label": "Helpers",
-    "type": "heading"
-  },
-  {
-    "id": "4:a",
-    "label": "trapFocus",
-    "route": "/react/trap-focus",
-    "tag": "",
-    "type": "route"
-  }
-]
